add nvidia-smi efficiency control example
nvidia-smi use sampleing technology to get the some core utilization percentage,
That is not the core used percentage. It is difficult to control the utilization of one core,
but you can modify the kernel launch frequency. It's creative to use recent history to dynamicly update
the latency, and it is a huge exchallenge to get a high accuracy model.

use:

```bash
cd src/nvsmi_control
make
```